Extract hasNoAliasingAccessInScope in MemRefUtils
