+++ mlir/lib/Dialect/MemRef/Utils/CMakeLists.txt
@@ -8,4 +8,5 @@ add_mlir_dialect_library(MLIRMemRefUtils
   MLIRIR
   MLIRAffineDialect
   MLIRArithUtils
+  MLIRSideEffectInterfaces
 )
